requirements.txt
alembic
click==7.0
coverage==4.2
Flask
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.6.1
itsdangerous==0.24
Mako==1.1.5
MarkupSafe==0.23
psycopg2
PyJWT==1.4.2
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.4.23
clock
PyCryptodome
bs4
lxml
fake_useragent
threading